# Docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Nifty100 Financial Intelligence API","description":"REST API for Nifty100 financial intelligence data.","version":"1.0.0"},"paths":{"/api/v1/companies":{"get":{"tags":["Companies"],"summary":"Get Companies","description":"Return a list of companies with optional filters.","operationId":"get_companies_api_v1_companies_get","parameters":[{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"market_cap_category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Market Cap Category"}},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}":{"get":{"tags":["Companies"],"summary":"Get Company","description":"Return the full profile of one company.","operationId":"get_company_api_v1_companies__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/profit-loss":{"get":{"tags":["Financials"],"summary":"Get Profit Loss","description":"Return profit and loss data.","operationId":"get_profit_loss_api_v1_companies__ticker__profit_loss_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Year"}},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/balance-sheet":{"get":{"tags":["Financials"],"summary":"Get Balance Sheet","description":"Return balance sheet data.","operationId":"get_balance_sheet_api_v1_companies__ticker__balance_sheet_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Year"}},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/cash-flow":{"get":{"tags":["Financials"],"summary":"Get Cash Flow","description":"Return cash flow data.","operationId":"get_cash_flow_api_v1_companies__ticker__cash_flow_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Year"}},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/ratios":{"get":{"tags":["Financials"],"summary":"Get Ratios","description":"Return financial ratios for a company.","operationId":"get_ratios_api_v1_companies__ticker__ratios_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/screener":{"get":{"tags":["Screener"],"summary":"Screen Companies","description":"Screen companies using financial and valuation filters.","operationId":"screen_companies_api_v1_screener_get","parameters":[{"name":"min_roe","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Roe"}},{"name":"max_de","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Max De"}},{"name":"min_fcf","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Fcf"}},{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"min_rev_cagr_5yr","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Rev Cagr 5Yr"}},{"name":"min_pat_cagr_5yr","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Pat Cagr 5Yr"}},{"name":"max_pe","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Max Pe"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sectors":{"get":{"tags":["Sectors"],"summary":"Get Sectors","description":"Return sector/peer-group summary information.","operationId":"get_sectors_api_v1_sectors_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/peers":{"get":{"tags":["Peers"],"summary":"Get Peer Groups","description":"Return all peer groups with company counts.","operationId":"get_peer_groups_api_v1_peers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/peers/{ticker}":{"get":{"tags":["Peers"],"summary":"Get Company Peers","description":"Return the peer group and peer companies for a ticker.","operationId":"get_company_peers_api_v1_peers__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/valuation":{"get":{"tags":["Valuation"],"summary":"Get Valuation","description":"Return valuation summary for all companies.","operationId":"get_valuation_api_v1_valuation_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/valuation/{ticker}":{"get":{"tags":["Valuation"],"summary":"Get Company Valuation","description":"Return valuation information for a single company.","operationId":"get_company_valuation_api_v1_valuation__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/portfolio":{"get":{"tags":["Portfolio"],"summary":"Get Portfolio","description":"Return portfolio summary for all companies.","operationId":"get_portfolio_api_v1_portfolio_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/documents":{"get":{"tags":["Documents"],"summary":"Get Documents","description":"Return document/annual-report records for all companies.","operationId":"get_documents_api_v1_documents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/documents/{ticker}":{"get":{"tags":["Documents"],"summary":"Get Company Documents","description":"Return documents for a specific company.","operationId":"get_company_documents_api_v1_documents__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/health":{"get":{"summary":"Health","description":"Return API and database health information.","operationId":"health_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","description":"Return basic API information.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}